# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API маркетплейса Клевер (микросервисная архитектура).\n- Auth Service       (HTTP :8001, gRPC :9001) — авторизация, профиль, пользователи\n- Support Service    (HTTP :8002)             — техподдержка\n- Catalog Service    (HTTP :8004, gRPC :9004) — объявления, категории, избранное, просмотры\n- Commerce Service   (HTTP :8006)             — корзина, чаты, заказы\nВсе запросы идут через API Gateway (nginx :8080).",
        "title": "Clover API",
        "contact": {},
        "version": "1.0"
    },
    "host": "clover-go.ru",
    "basePath": "/api/v1",
    "paths": {
...
        "/support/tickets/{id}/messages": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает все сообщения чата обращения. Доступно автору обращения,\nа также пользователям с ролью support/admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Получить сообщения обращения",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID обращения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "список сообщений",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.MessageResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request: неверный ID или нет доступа",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized: пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error: ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Отправляет сообщение в чат обращения. Доступно автору обращения,\nа также пользователям с ролью support/admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Отправить сообщение в чат обращения",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID обращения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Текст сообщения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "сообщение отправлено",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "bad request: пустой текст, неверный ID или нет доступа",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized: пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error: ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.MessageResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "ticket_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SendMessageRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT session token",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        },
        "CsrfCookieAuth": {
            "description": "CSRF token stored in cookies",
            "type": "apiKey",
            "name": "csrf_token",
            "in": "cookie"
        },
        "CsrfHeaderAuth": {
            "description": "CSRF token required in header",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}